+++ CHANGELOG.md
@@ -7,8 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `JSONDeserializer`: added deserialization of Kotlin `object`s (serialization already worked)
 - `JSONSerializer`: improved serialization of Java objects
 - `JSONDeserializer`: improved deserialization of Java objects
-- `JSONSerializer`: added serialization of `java.time.Duration`
-- `JSONDeserializer`: added deserialization of `java.time.Duration`
+- `JSONSerializer`: added serialization of `java.time.Duration`, `java.time.Period`
+- `JSONDeserializer`: added deserialization of `java.time.Duration`, `java.time.Period`
 - `JSONConfig`: added `readBufferSize`
 
 ## [0.6] - 2019-05-23